move helpers to libbeat, add some godocs
