+++ metricbeat/module/docker/config.go
@@ -17,6 +17,7 @@
 
 package docker
 
+// Config contains the config needed for the docker module
 type Config struct {
 	TLS   *TLSConfig `config:"ssl"`
 	DeDot bool       `config:"labels.dedot"`
@@ -29,6 +30,7 @@ func DefaultConfig() Config {
 	}
 }
 
+// TLSConfig contains TLS config
 type TLSConfig struct {
 	Enabled     *bool  `config:"enabled"`
 	CA          string `config:"certificate_authority"`
@@ -36,6 +38,7 @@ type TLSConfig struct {
 	Key         string `config:"key"`
 }
 
+// IsEnabled return true if TLS is enabled
 func (c *TLSConfig) IsEnabled() bool {
 	return c != nil && (c.Enabled == nil || *c.Enabled)
 }
