+++ metricbeat/module/docker/helper.go
@@ -18,14 +18,13 @@
 package docker
 
 import (
-	"strings"
-
 	"github.com/docker/docker/api/types"
 
 	"github.com/elastic/beats/libbeat/common"
-	"github.com/elastic/beats/libbeat/common/safemapstr"
+	helpers "github.com/elastic/beats/libbeat/common/docker"
 )
 
+// Container is a struct representation of a container
 type Container struct {
 	ID     string
 	Name   string
@@ -33,6 +32,7 @@ type Container struct {
 	Labels common.MapStr
 }
 
+// ToMapStr converts a container struct to a MapStrs
 func (c *Container) ToMapStr() common.MapStr {
 	m := common.MapStr{
 		"container": common.MapStr{
@@ -57,41 +57,8 @@ func (c *Container) ToMapStr() common.MapStr {
 func NewContainer(container *types.Container, dedot bool) *Container {
 	return &Container{
 		ID:     container.ID,
-		Name:   ExtractContainerName(container.Names),
-		Labels: DeDotLabels(container.Labels, dedot),
+		Name:   helpers.ExtractContainerName(container.Names),
+		Labels: helpers.DeDotLabels(container.Labels, dedot),
 		Image:  container.Image,
 	}
-}
-
-func ExtractContainerName(names []string) string {
-	output := names[0]
-
-	if len(names) > 1 {
-		for _, name := range names {
-			if strings.Count(output, "/") > strings.Count(name, "/") {
-				output = name
-			}
-		}
-	}
-	return strings.Trim(output, "/")
-}
-
-// DeDotLabels returns a new common.MapStr containing a copy of the labels
-// where the dots have been converted into nested structure, avoiding
-// possible mapping errors
-func DeDotLabels(labels map[string]string, dedot bool) common.MapStr {
-	outputLabels := common.MapStr{}
-	for k, v := range labels {
-		if dedot {
-			// This is necessary so that ES does not interpret '.' fields as new
-			// nested JSON objects, and also makes this compatible with ES 2.x.
-			label := common.DeDot(k)
-			outputLabels.Put(label, v)
-		} else {
-			// If we don't dedot we ensure there are no mapping errors with safemapstr
-			safemapstr.Put(outputLabels, k, v)
-		}
-	}
-
-	return outputLabels
 }
